requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==20.1.0             # via pytest
click==7.1.2              # via flask
coverage==5.2.1           # via -r requirements.in
flask==1.1.2              # via -r requirements.in
importlib-metadata==1.7.0  # via pluggy, pytest
iniconfig==1.0.1          # via pytest
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyjwt==1.7.1              # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest==6.0.1             # via -r requirements.in
six==1.15.0               # via packaging
sqlalchemy==1.3.18        # via -r requirements.in
toml==0.10.1              # via pytest
werkzeug==1.0.1           # via flask
zipp==3.1.0               # via importlib-metadata
